audio_chatbot/audio_chatbot/requirements.txt
django==4.0.0
channels==3.0.4
google-generativeai>=0.3.0
SpeechRecognition>=3.10.0
whitenoise>=6.0.0
psycopg2-binary>=2.9.0
dj-database-url>=2.1.0
python-dotenv>=1.0.0
pyaudio>=0.2.13